requirements.txt
PyMuPDF>=1.24
openai-whisper
pdf2image
Pillow
requests
duckduckgo-search
